@@ -33,32 +33,32 @@
         }
     ],
     "require": {
-        "php": "^8.1",
-        "alexusmai/laravel-file-manager": "^3.0",
+        "php": "^8.2",
+        "alexusmai/laravel-file-manager": "^3.2",
         "barryvdh/laravel-snappy": "^1.0",
-        "bugsnag/bugsnag-laravel": "^2.26",
+        "bugsnag/bugsnag-laravel": "^2.27",
         "doctrine/dbal": "^3.5",
-        "eduardokum/laravel-mail-auto-embed": "^2.10",
-        "fedeisas/laravel-mail-css-inliner": "^5.2",
+        "eduardokum/laravel-mail-auto-embed": "^2.11",
+        "fedeisas/laravel-mail-css-inliner": "^5.3",
         "fennb/phirehose": "^1.0",
         "google/recaptcha": "^1.2",
-        "graham-campbell/markdown": "^15.0",
-        "guzzlehttp/guzzle": "^7.2",
+        "graham-campbell/markdown": "^15.2",
+        "guzzlehttp/guzzle": "^7.8",
         "h4cc/wkhtmltoimage-amd64": "0.12.x",
         "h4cc/wkhtmltopdf-amd64": "dev-master#4144bccbfaa40e8150a763e2d55dfedce9d13909",
         "intervention/image": "^2.6",
-        "jdavidbakr/mail-tracker": "^7.0",
+        "jdavidbakr/mail-tracker": "^7.15",
         "jenssegers/agent": "^2.6",
-        "knplabs/knp-snappy": "^1.4",
+        "knplabs/knp-snappy": "^1.5",
         "kordy/ticketit": "dev-master",
-        "kriswallsmith/buzz": "^1.2",
-        "laravel/cashier": "^14.7",
-        "laravel/dusk": "^7.4",
-        "laravel/framework": "^10.48",
-        "laravel/socialite": "^5.6",
-        "laravel/telescope": "^4.12",
-        "laravel/tinker": "^2.8",
-        "laravel/ui": "^4.2",
+        "kriswallsmith/buzz": "^1.3",
+        "laravel/cashier": "^15.2",
+        "laravel/dusk": "^8.0",
+        "laravel/framework": "^11.41",
+        "laravel/socialite": "^5.11",
+        "laravel/telescope": "^5.0",
+        "laravel/tinker": "^2.9",
+        "laravel/ui": "^4.4",
         "laravelcollective/html": "^6.4",
         "league/flysystem": "^3.0",
         "league/flysystem-aws-s3-v3": "^3.0",
@@ -70,23 +70,22 @@
         "paquettg/php-html-parser": "^2.2",
         "phpmd/phpmd": "^2.6",
         "predis/predis": "^1.1",
-        "rap2hpoutre/fast-excel": "^5.1",
-        "shanmuga/laravel-entrust": "^4.0",
+        "rap2hpoutre/fast-excel": "^5.4",
+        "shanmuga/laravel-entrust": "^5.0",
         "socialiteproviders/linkedin": "^4.1",
-        "spatie/browsershot": "^3.52",
-        "spatie/laravel-activitylog": "^4.7",
-        "spatie/laravel-referer": "^1.8",
+        "spatie/browsershot": "^3.60",
+        "spatie/laravel-activitylog": "^4.8",
+        "spatie/laravel-referer": "^1.9",
         "squizlabs/php_codesniffer": "^3.6",
         "stripe/stripe-php": "^7.94"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.8",
-        "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^6.3",
-        "phpunit/phpunit": "^9.5.10",
-        "fakerphp/faker": "^1.9.1",
-        "diglabby/laravel-find-missing-translations": "dev-master",
-        "spatie/laravel-ignition": "^2.0"
+        "barryvdh/laravel-debugbar": "^3.10",
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.0",
+        "phpunit/phpunit": "^10.5",
+        "fakerphp/faker": "^1.23",
+        "diglabby/laravel-find-missing-translations": "dev-master"
     },
     "extra": {
         "laravel": {
