@@ -15,12 +15,15 @@ class OrgAdminProp extends Model
 
     const UPDATED_AT = 'updateDate';
 
-    protected $casts = [
-        'createDate' => 'datetime',
-        'updateDate' => 'datetime',
-    ];
-
     protected $fillable = ['orgID', 'propID', 'value'];
+
+    protected function casts(): array
+    {
+        return [
+            'createDate' => 'datetime',
+            'updateDate' => 'datetime',
+        ];
+    }
 
     protected function setKeysForSaveQuery($query)
     {
